Move APIs to Module surfaces
Rust doesn't lend itself well to the same multi-struct architecture
used by, for example, Resend's Go client. Instead, this moves the
individual APIs to the surface of each module and requires a reference
to the shared client be passed in, inverting control and giving it
back to the consumer of the library.